+++ Sistema.Web/Models/Ventas/VentaDetalle/ActualizarViewModel.cs
@@ -10,7 +10,7 @@ namespace Sistema.Web.Models.Ventas.VentaDetalle
         public int ProId { get; set; }
         public decimal VedPrecio { get; set; }
         public decimal VedDescuento { get; set; }
-        public float VedCantidad { get; set; }
+        public double VedCantidad { get; set; }
 
     }
 }
