+++ Sistema.Web/Models/Ventas/VentaDetalle/VentaDetalleViewModel.cs
@@ -7,7 +7,7 @@ namespace Sistema.Web.Models.Ventas.VentaDetalle
         public int ProId { get; set; }
         public decimal VedPrecio { get; set; }
         public decimal VedDescuento { get; set; }
-        public float VedCantidad { get; set; }
+        public double VedCantidad { get; set; }
         public string Venta { get; set; }
         public string Producto { get; set; }
     }
